refactor pydantic tests
