+++ tests/unit/test_contrib/test_pydantic/test_dto.py
@@ -1,0 +1,33 @@
+from __future__ import annotations
+
+from pydantic import BaseModel
+
+from litestar import Request, post
+from litestar.contrib.pydantic import PydanticDTO, _model_dump_json
+from litestar.dto import DTOConfig
+from litestar.testing import create_test_client
+
+
+def test_schema_required_fields_with_pydantic_dto(use_experimental_dto_backend: bool) -> None:
+    class PydanticUser(BaseModel):
+        age: int
+        name: str
+
+    class UserDTO(PydanticDTO[PydanticUser]):
+        config = DTOConfig(experimental_codegen_backend=use_experimental_dto_backend)
+
+    @post(dto=UserDTO, return_dto=None, signature_types=[PydanticUser])
+    def handler(data: PydanticUser, request: Request) -> dict:
+        schema = request.app.openapi_schema
+        return schema.to_schema()
+
+    with create_test_client(handler) as client:
+        data = PydanticUser(name="A", age=10)
+        headers = {"Content-Type": "application/json; charset=utf-8"}
+        received = client.post(
+            "/",
+            content=_model_dump_json(data),
+            headers=headers,
+        )
+        required = next(iter(received.json()["components"]["schemas"].values()))["required"]
+        assert len(required) == 2
