+++ tests/unit/test_contrib/test_pydantic/test_integration.py
@@ -1,13 +1,15 @@
 from __future__ import annotations
 
-from typing import Any
+from typing import Any, List
 
 from pydantic import VERSION, BaseModel, Field
 
 from litestar import post
 from litestar.contrib.pydantic.pydantic_dto_factory import PydanticDTO
+from litestar.params import Parameter
+from litestar.status_codes import HTTP_400_BAD_REQUEST
 from litestar.testing import create_test_client
-from tests.models import PydanticPerson
+from tests.unit.test_contrib.test_pydantic.models import PydanticPerson
 
 
 def test_pydantic_validation_error_raises_400() -> None:
@@ -67,3 +69,62 @@ def test_default_handling_of_pydantic_errors() -> None:
         extra = response.json().get("extra")
         assert extra is not None
         assert 3 if len(extra) == VERSION.startswith("1") else 4
+
+
+def test_signature_model_invalid_input() -> None:
+    class OtherChild(BaseModel):
+        val: List[int]  # noqa: UP006
+
+    class Child(BaseModel):
+        val: int
+        other_val: int
+
+    class Parent(BaseModel):
+        child: Child
+        other_child: OtherChild
+
+    @post("/")
+    def test(
+        data: Parent,
+        int_param: int,
+        length_param: str = Parameter(min_length=2),
+        int_header: int = Parameter(header="X-SOME-INT"),
+        int_cookie: int = Parameter(cookie="int-cookie"),
+    ) -> None:
+        ...
+
+    with create_test_client(route_handlers=[test], signature_types=[Parent]) as client:
+        client.cookies.update({"int-cookie": "cookie"})
+        response = client.post(
+            "/",
+            json={"child": {"val": "a", "other_val": "b"}, "other_child": {"val": [1, "c"]}},
+            params={"int_param": "param", "length_param": "d"},
+            headers={"X-SOME-INT": "header"},
+        )
+
+        assert response.status_code == HTTP_400_BAD_REQUEST
+
+        data = response.json()
+
+        assert data
+        if VERSION.startswith("1"):
+            assert data["extra"] == [
+                {"key": "child.val", "message": "value is not a valid integer"},
+                {"key": "child.other_val", "message": "value is not a valid integer"},
+                {"key": "other_child.val.1", "message": "value is not a valid integer"},
+            ]
+        else:
+            assert data["extra"] == [
+                {
+                    "message": "Input should be a valid integer, unable to parse string as an integer",
+                    "key": "child.val",
+                },
+                {
+                    "message": "Input should be a valid integer, unable to parse string as an integer",
+                    "key": "child.other_val",
+                },
+                {
+                    "message": "Input should be a valid integer, unable to parse string as an integer",
+                    "key": "other_child.val.1",
+                },
+            ]
