+++ tests/unit/test_kwargs/test_cookie_params.py
@@ -1,7 +1,6 @@
 from typing import Optional, Type
 
 import pytest
-from pydantic.fields import FieldInfo
 
 from litestar import get
 from litestar.params import Parameter
@@ -12,22 +11,13 @@ from litestar.testing import create_test_client
 @pytest.mark.parametrize(
     "t_type,param_dict,param,expected_code",
     [
-        #     str,
-        #     HTTP_400_BAD_REQUEST,
-        # ),
-        (
-            Optional[str],
-            {},
-            Parameter(cookie="special-cookie", min_length=1, max_length=2, required=False),
-            HTTP_200_OK,
-        ),
         (int, {"special-cookie": "123"}, Parameter(cookie="special-cookie", ge=100, le=201), HTTP_200_OK),
         (int, {"special-cookie": "123"}, Parameter(cookie="special-cookie", ge=100, le=120), HTTP_400_BAD_REQUEST),
         (int, {}, Parameter(cookie="special-cookie", ge=100, le=120), HTTP_400_BAD_REQUEST),
         (Optional[int], {}, Parameter(cookie="special-cookie", ge=100, le=120, required=False), HTTP_200_OK),
     ],
 )
-def test_cookie_params(t_type: Type, param_dict: dict, param: FieldInfo, expected_code: int) -> None:
+def test_cookie_params(t_type: Type, param_dict: dict, param: Parameter, expected_code: int) -> None:
     test_path = "/test"
 
     @get(path=test_path)
