+++ tests/unit/test_kwargs/test_json_data.py
@@ -1,5 +1,6 @@
+from dataclasses import asdict
+
 from litestar import post
-from litestar.contrib.pydantic import _model_dump
 from litestar.params import Body
 from litestar.status_codes import HTTP_201_CREATED
 from litestar.testing import create_test_client
@@ -13,9 +14,7 @@ def test_request_body_json() -> None:
         assert isinstance(data, Form)
 
     with create_test_client(test_method) as client:
-        response = client.post(
-            "/test", json=_model_dump(Form(name="Moishe Zuchmir", age=30, programmer=True, value="100"))
-        )
+        response = client.post("/test", json=asdict(Form(name="Moishe Zuchmir", age=30, programmer=True, value="100")))
         assert response.status_code == HTTP_201_CREATED
