+++ tests/unit/test_kwargs/test_multipart_data.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 
 from collections import defaultdict
-from dataclasses import dataclass
+from dataclasses import asdict, dataclass
 from os import path
 from os.path import dirname, join, realpath
 from pathlib import Path
@@ -10,13 +10,9 @@ from typing import Any, DefaultDict, Dict, List, Optional
 
 import msgspec
 import pytest
-from attr import define, field
-from attr.validators import ge, instance_of, lt
-from pydantic import BaseConfig, BaseModel, ConfigDict, Field
 from typing_extensions import Annotated
 
 from litestar import Request, post
-from litestar.contrib.pydantic import _model_dump
 from litestar.datastructures.upload_file import UploadFile
 from litestar.enums import RequestEncodingType
 from litestar.params import Body
@@ -26,13 +22,11 @@ from litestar.testing import create_test_client
 from . import Form
 
 
-class FormData(BaseModel):
+@dataclass
+class FormData:
     name: UploadFile
     age: UploadFile
     programmer: UploadFile
-
-    class Config(BaseConfig):
-        arbitrary_types_allowed = True
 
 
 @post("/form")
@@ -93,7 +87,7 @@ async def form_with_headers_handler(request: Request) -> Dict[str, Any]:
 @pytest.mark.parametrize("t_type", [FormData, Dict[str, UploadFile], List[UploadFile], UploadFile])
 def test_request_body_multi_part(t_type: type) -> None:
     test_path = "/test"
-    data = _model_dump(Form(name="Moishe Zuchmir", age=30, programmer=True, value="100"))
+    data = asdict(Form(name="Moishe Zuchmir", age=30, programmer=True, value="100"))
 
     @post(path=test_path, signature_namespace={"t_type": t_type})
     def test_method(data: Annotated[t_type, Body(media_type=RequestEncodingType.MULTI_PART)]) -> None:  # type: ignore
@@ -493,31 +487,15 @@ def test_multipart_handling_of_none_values() -> None:
         assert response.status_code == HTTP_201_CREATED
 
 
-MAX_INT_POSTGRES = 10
-
-
-@define
-class AddProductFormAttrs:
-    name: str
-    amount: int = field(validator=[instance_of(int), ge(1), lt(MAX_INT_POSTGRES)])
-
-
-class AddProductFormPydantic(BaseModel):
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-    name: str
-    amount: int = Field(ge=1, lt=MAX_INT_POSTGRES)
-
-
 class AddProductFormMsgspec(msgspec.Struct):
     name: str
-    amount: Annotated[int, msgspec.Meta(lt=MAX_INT_POSTGRES, ge=1)]
+    amount: Annotated[int, msgspec.Meta(lt=10, ge=1)]
 
 
-@pytest.mark.parametrize("form_object", [AddProductFormMsgspec, AddProductFormPydantic, AddProductFormAttrs])
 @pytest.mark.parametrize("form_type", [RequestEncodingType.URL_ENCODED, RequestEncodingType.MULTI_PART])
-def test_multipart_and_url_encoded_behave_the_same(form_object, form_type) -> None:  # type: ignore[no-untyped-def]
-    @post(path="/form", signature_namespace={"form_object": form_object, "form_type": form_type})
-    async def form_(request: Request, data: Annotated[form_object, Body(media_type=form_type)]) -> int:
+def test_multipart_and_url_encoded_behave_the_same(form_type) -> None:  # type: ignore[no-untyped-def]
+    @post(path="/form", signature_namespace={"form_object": AddProductFormMsgspec, "form_type": form_type})
+    async def form_(request: Request, data: Annotated[AddProductFormMsgspec, Body(media_type=form_type)]) -> int:
         assert isinstance(data.name, str)
         return data.amount  # type: ignore[no-any-return]
