+++ tests/unit/test_kwargs/test_url_encoded_data.py
@@ -1,7 +1,7 @@
+from dataclasses import asdict
 from typing import Optional
 
 from litestar import post
-from litestar.contrib.pydantic import _model_dump
 from litestar.enums import RequestEncodingType
 from litestar.params import Body
 from litestar.status_codes import HTTP_201_CREATED
@@ -16,9 +16,7 @@ def test_request_body_url_encoded() -> None:
         assert isinstance(data, Form)
 
     with create_test_client(test_method) as client:
-        response = client.post(
-            "/test", data=_model_dump(Form(name="Moishe Zuchmir", age=30, programmer=True, value="100"))
-        )
+        response = client.post("/test", data=asdict(Form(name="Moishe Zuchmir", age=30, programmer=True, value="100")))
         assert response.status_code == HTTP_201_CREATED
