+++ tests/unit/test_openapi/test_config.py
@@ -1,21 +1,16 @@
-from sys import version_info
 from typing import TYPE_CHECKING, Any
 
 import pytest
-from pydantic import BaseModel, Field
 
-from litestar import Litestar, get, post
-from litestar.contrib.pydantic import PydanticPlugin
+from litestar import Litestar, get
 from litestar.exceptions import ImproperlyConfiguredException
 from litestar.openapi.config import OpenAPIConfig
 from litestar.openapi.spec import Components, Example, OpenAPIHeader, OpenAPIType, Schema
-from litestar.testing import TestClient
 
 if TYPE_CHECKING:
     from litestar.handlers.http_handlers import HTTPRouteHandler
 
 
-@pytest.mark.skipif(version_info < (3, 10), reason="pydantic serialization differences in lower python versions")
 def test_merged_components_correct() -> None:
     components_one = Components(headers={"one": OpenAPIHeader()}, schemas={"test": Schema(type=OpenAPIType.STRING)})
     components_two = Components(headers={"two": OpenAPIHeader()})
@@ -47,80 +42,6 @@ def test_merged_components_correct() -> None:
             },
         },
     }
-
-
-def test_by_alias() -> None:
-    class RequestWithAlias(BaseModel):
-        first: str = Field(alias="second")
-
-    class ResponseWithAlias(BaseModel):
-        first: str = Field(alias="second")
-
-    @post("/")
-    def handler(data: RequestWithAlias) -> ResponseWithAlias:
-        return ResponseWithAlias(second=data.first)
-
-    app = Litestar(route_handlers=[handler], openapi_config=OpenAPIConfig(title="my title", version="1.0.0"))
-
-    assert app.openapi_schema
-    schemas = app.openapi_schema.to_schema()["components"]["schemas"]
-    request_key = "second"
-    assert schemas["RequestWithAlias"] == {
-        "properties": {request_key: {"type": "string"}},
-        "type": "object",
-        "required": [request_key],
-        "title": "RequestWithAlias",
-    }
-    response_key = "first"
-    assert schemas["ResponseWithAlias"] == {
-        "properties": {response_key: {"type": "string"}},
-        "type": "object",
-        "required": [response_key],
-        "title": "ResponseWithAlias",
-    }
-
-    with TestClient(app) as client:
-        response = client.post("/", json={request_key: "foo"})
-        assert response.json() == {response_key: "foo"}
-
-
-def test_pydantic_plugin_override_by_alias() -> None:
-    class RequestWithAlias(BaseModel):
-        first: str = Field(alias="second")
-
-    class ResponseWithAlias(BaseModel):
-        first: str = Field(alias="second")
-
-    @post("/")
-    def handler(data: RequestWithAlias) -> ResponseWithAlias:
-        return ResponseWithAlias(second=data.first)
-
-    app = Litestar(
-        route_handlers=[handler],
-        openapi_config=OpenAPIConfig(title="my title", version="1.0.0"),
-        plugins=[PydanticPlugin(prefer_alias=True)],
-    )
-
-    assert app.openapi_schema
-    schemas = app.openapi_schema.to_schema()["components"]["schemas"]
-    request_key = "second"
-    assert schemas["RequestWithAlias"] == {
-        "properties": {request_key: {"type": "string"}},
-        "type": "object",
-        "required": [request_key],
-        "title": "RequestWithAlias",
-    }
-    response_key = "second"
-    assert schemas["ResponseWithAlias"] == {
-        "properties": {response_key: {"type": "string"}},
-        "type": "object",
-        "required": [response_key],
-        "title": "ResponseWithAlias",
-    }
-
-    with TestClient(app) as client:
-        response = client.post("/", json={request_key: "foo"})
-        assert response.json() == {response_key: "foo"}
 
 
 def test_allows_customization_of_operation_id_creator() -> None:
