+++ tests/unit/test_signature/test_validation.py
@@ -3,12 +3,10 @@ from typing import List, Optional
 
 import pytest
 from attr import define
-from pydantic import VERSION, BaseModel
 from typing_extensions import TypedDict
 
 from litestar import get, post
 from litestar._signature import SignatureModel
-from litestar.contrib.pydantic import PydanticInitPlugin
 from litestar.di import Provide
 from litestar.exceptions import ImproperlyConfiguredException, ValidationException
 from litestar.params import Dependency, Parameter
@@ -102,18 +100,23 @@ def test_client_backend_error_precedence_over_server_error() -> None:
 
 
 def test_validation_error_exception_key() -> None:
-    class OtherChild(BaseModel):
+    from dataclasses import dataclass
+
+    @dataclass
+    class OtherChild:
         val: List[int]
 
-    class Child(BaseModel):
+    @dataclass
+    class Child:
         val: int
         other_val: int
 
-    class Parent(BaseModel):
+    @dataclass
+    class Parent:
         child: Child
         other_child: OtherChild
 
-    @get("/", type_decoders=PydanticInitPlugin.decoders())
+    @get("/")
     def handler(data: Parent) -> None:
         pass
 
@@ -131,68 +134,7 @@ def test_validation_error_exception_key() -> None:
         )
 
     assert isinstance(exc_info.value.extra, list)
-    assert exc_info.value.extra[0]["key"] == "child.val"
-    assert exc_info.value.extra[1]["key"] == "child.other_val"
-    assert exc_info.value.extra[2]["key"] == "other_child.val"
-
-
-def test_invalid_input_pydantic() -> None:
-    class OtherChild(BaseModel):
-        val: List[int]
-
-    class Child(BaseModel):
-        val: int
-        other_val: int
-
-    class Parent(BaseModel):
-        child: Child
-        other_child: OtherChild
-
-    @post("/")
-    def test(
-        data: Parent,
-        int_param: int,
-        length_param: str = Parameter(min_length=2),
-        int_header: int = Parameter(header="X-SOME-INT"),
-        int_cookie: int = Parameter(cookie="int-cookie"),
-    ) -> None:
-        ...
-
-    with create_test_client(route_handlers=[test]) as client:
-        client.cookies.update({"int-cookie": "cookie"})
-        response = client.post(
-            "/",
-            json={"child": {"val": "a", "other_val": "b"}, "other_child": {"val": [1, "c"]}},
-            params={"int_param": "param", "length_param": "d"},
-            headers={"X-SOME-INT": "header"},
-        )
-
-        assert response.status_code == HTTP_400_BAD_REQUEST
-
-        data = response.json()
-
-        assert data
-        if VERSION.startswith("1"):
-            assert data["extra"] == [
-                {"key": "child.val", "message": "value is not a valid integer"},
-                {"key": "child.other_val", "message": "value is not a valid integer"},
-                {"key": "other_child.val.1", "message": "value is not a valid integer"},
-            ]
-        else:
-            assert data["extra"] == [
-                {
-                    "message": "Input should be a valid integer, unable to parse string as an integer",
-                    "key": "child.val",
-                },
-                {
-                    "message": "Input should be a valid integer, unable to parse string as an integer",
-                    "key": "child.other_val",
-                },
-                {
-                    "message": "Input should be a valid integer, unable to parse string as an integer",
-                    "key": "other_child.val.1",
-                },
-            ]
+    assert exc_info.value.extra[0]["key"] == "child"
 
 
 def test_invalid_input_attrs() -> None:
